pyproject.toml
[project]
name = "async-standup"
version = "0.1.0"
description = "AsyncStandup emotion detection demo using Smallest.ai Pulse API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "pydub>=0.25.1",
    "requests>=2.32.5",
    "smallestai>=4.3.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
